+++ .idea/workspace.xml
@@ -4,7 +4,9 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="0330bf4e-987b-4d9c-9f25-17436b0c41d5" name="Changes" comment="" />
+    <list default="true" id="0330bf4e-987b-4d9c-9f25-17436b0c41d5" name="Changes" comment="">
+      <change beforePath="$PROJECT_DIR$/src/sudokugame/SudokuFrame.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/sudokugame/SudokuFrame.java" afterDir="false" />
+    </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
@@ -62,8 +64,8 @@
     </key>
   </component>
   <component name="RunManager">
-    <configuration name="SudokuGame" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
-      <option name="MAIN_CLASS_NAME" value="sudokugame.SudokuGame" />
+    <configuration name="SudokuLauncher" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
+      <option name="MAIN_CLASS_NAME" value="sudokugame.SudokuLauncher" />
       <module name="MySudokuGame" />
       <extension name="coverage">
         <pattern>
@@ -77,7 +79,7 @@
     </configuration>
     <recent_temporary>
       <list>
-        <item itemvalue="Application.SudokuGame" />
+        <item itemvalue="Application.SudokuLauncher" />
       </list>
     </recent_temporary>
   </component>
