--- src/sudokugame/SudokuGame.java
@@ -1,8 +1,0 @@
-package sudokugame;
-
-public class SudokuGame {
-
-    public static void main(String[] args) {
-
-    }
-}
