+++ src/sudokugame/SudokuMainMenu.java
@@ -110,24 +110,4 @@ public class SudokuMainMenu extends JFrame {
         new SudokuFrame(difficulty).setVisible(true);
         this.dispose();  // Close the main menu
     }
-
-    public static void main(String[] args) {
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException | InstantiationException | IllegalAccessException |
-                 UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(SudokuMainMenu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new SudokuMainMenu().setVisible(true);
-            }
-        });
-    }
 }
